Move from Travis CI to Github Actions
This is as far as I could get without having credentials to the Docker
registry.  This build fails at `docker login` since it doesn't have
those creds loaded into environment variables in github.

The Travis CI should continue to work until the `.travis.yml` file is
removed.  The `Makefile` recipes are still there as well and can be
deleted once we're ready to decommission the Travis build.

To finish migrating to GH Actions you'll need to populate the
environment variables in Github with the username/password and registry.